POST player/session fixed (request types)

+++ EntertainmentPortal/EP.Sudoku.Data/Models/CellDb.cs
@@ -16,18 +16,21 @@ namespace EP.Sudoku.Data.Models
         /// Is used to denote a value (number) from 0 till 9 for each cell filling up (DbInfo).
         /// </summary>      
         public int Value { get; set; }
-        
+
         /// <summary>
-        /// Is used to denote cell's coordinates on the gameboard(X and Y values are expected to be used). 
+        /// Is used to denote cell's coordinates on the gameboard(X and Y values are expected to be used) (DbInfo). 
         /// </summary>
         public int X { get; set; }
 
         /// <summary>
-        /// Is used to denote cell's coordinates on the gameboard (X and Y values are expected to be used). 
+        /// Is used to denote cell's coordinates on the gameboard (X and Y values are expected to be used) (DbInfo). 
         /// </summary>
         public int Y { get; set; }
 
-        public int SessionDbId { get; set; }
+        /// <summary>
+        /// Is used to provide coinsaidence between a game session and a cell of the gameboard (DbInfo).
+        /// </summary> 
+        public long SessionDbId { get; set; }
     }
 }
